% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_mcnaqc.R
\name{All_segments_mutations}
\alias{All_segments_mutations}
\title{Extract mutations from common and private segments from multi_CNAqc object.}
\usage{
All_segments_mutations(
  m_cnaqc_obj,
  sample = "all",
  cna = c("clonal", "subclonal"),
  type = c("SNV", "indel")
)
}
\arguments{
\item{m_cnaqc_obj}{multi_CNAqc object (must have previously run \code{"multisample_init"})}

\item{sample}{desired samples for which you want to extract the mutations. Default \code{"all"}, extracts mutations from all samples
in the multi_CNAqc object.}

\item{cna}{\code{"clonal"} for clonal CNAs, \code{"subclonal"} for subclonal CNAs.}

\item{type}{\code{"SNV"} for single-nucleotide variants, \code{"indel"} for insertion-deletions.}
}
\value{
a named list of tibbles with all the mutations.
}
\description{
This function extracts the all the mutations in CNAqc format of desired samples from a multi_CNAqc object, 
merging together those that belong to common segments and those private to the sample.
}
\examples{
 
all_segments_mut = All_segments_mutations(example_multisample, 
                                sample = "all", 
                                cna = c("clonal", "subclonal"), 
                                type = c("SNV", "indel"))
print(all_segments_mut)
}
