% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_drivers.R
\name{annotate_drivers}
\alias{annotate_drivers}
\title{Title}
\usage{
annotate_drivers(
  x,
  drivers_list = CNAqc::DriverDBv3,
  function_entries = c("exonic", "Exonic"),
  function_column = "ANNOVAR_FUNCTION",
  gene_column = "GENE"
)
}
\arguments{
\item{x}{}

\item{drivers_list}{}

\item{function_column}{}

\item{gene_column}{}
}
\value{

}
\description{
Title
}
