% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{annotate_variants}
\alias{annotate_variants}
\title{Annotate variant in a CNAqc object}
\usage{
annotate_variants(
  x,
  ref = "hg19",
  driver_list = CNAqc::intogen_drivers,
  polyphen = FALSE,
  filter_tumor_type = NULL
)
}
\arguments{
\item{x}{a data.frame with coloumns c("chr", "from", "to").}

\item{driver_list}{a one coloumn data.frame with gene symbols}

\item{polyphen}{whete to annotate polyphen scores}

\item{filter_tumor_type}{filter intogen drivers by tumor type}
}
\value{
a data.frame with additiona coloumns `location` with the position of the variant in the genome (coding, intron, threeUTR, ...),
`consequence` with the consequence of coding mutations (synonymous, nonsynonymous, ...), `is_driver` a bool that indicate if the gene is a driver,
`gene_symbol` for the annotated corresponding gene symbol (if the variant is in a gene) and `driver_label` with the driver label written as
gene_refAA->varAA (is NA in case `is_driver` is FALSE).

The `chr` coloumn int the input should contain the chr prefix before the chromosome number. The annotation process is based on the package
\href{https://bioconductor.org/packages/release/bioc/html/VariantAnnotation.html}{VariantAnnotation} check it for more information.
}
\description{
The function provides an easy and fast way to annotate SNVs, in particular we annotate the locations of the different mutations, the consequences for
those in coding regions and we compare them to a set of known drivers.
}
\examples{

library(CNAqc)

data('example_dataset_CNAqc', package = 'CNAqc')

snvs <- example_dataset_CNAqc$snvs

snvs_annotated <- annotate_variants(snvs)
}
