% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vep.R
\name{augment_with_vep}
\alias{augment_with_vep}
\title{Import VEP annotations.}
\usage{
augment_with_vep(x, vep, columns = NULL)
}
\arguments{
\item{x}{A CNAqc object, or a dataframe of mutations in the input format
for CNAqc.}

\item{maf}{The file MAF associated containing the annotations in MAF format.}
}
\value{

}
\description{
This function imports
\href{https://www.ensembl.org/info/docs/tools/vep/index.html}{Ensembl Variant Effect Predictor (VEP)}
annotations into a CNAqc object, or to a mutations
dataframe in the format ready for CNAqc. Therefore the expected uses are

\itemize{
 \item{"[Option 1]"}{Mutations + CNA + Purity -> CNAqc -> VEP annotation}
 \item{"[Option 2]"}{Mutations -> VEP annotation -> Mutations (with VEP) + CNA + Purity -> CNAqc}
}

VEP annotations should have been created using the scripts discussed at
the
\href{https://github.com/caravagnalab/CNAqc}{CNAqc webpage}
utility. At this point, if one has created the input data for CNAqc from the
original VCF file, the VEP can be added to the CNAqc object (Option 1), or to
the available mutations before creating the CNAqc obejct (Option 2).

VEP annotations and input mutations are matched by exact genomic coordinates.
}
\examples{
# Example with a CNAqc input object
if(FALSE)
{
   # Create your CNAqc object (omissis here) from an original "file.vcf"
   x = init(mutations = ..., cna = ..., purity = ...)

   # Offline, create your MAF annotations as file "file_vcf.maf" from "file.vcf"
   # vcf2maf file.vcf .... file_vcf.maf

   # Import into R/CNAqc
   x = augment_with_maf(x, maf = "file_vcf.maf")

   # check they are in (there should be many columns with "MAF." prefix)
   x \%>\% Mutations \%>\% colnames
}
}
\seealso{
function \code{\link{as_maftools_cohort}} to convert multiple CNAqc
objects with MAF annotations into a single MAF cohort;  package
[maftools](https://bioconductor.org/packages/release/bioc/html/maftools.html) to
summarize, analyze and visualize MAF Files; the utility [vcf2maf](https://github.com/mskcc/vcf2maf)
to create MAF files from VCFs, using the [Ensembl Variant Effect Predictor (VEP)](https://www.ensembl.org/info/docs/tools/vep/index.html)
utility.
}
