% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_tolerance.R
\name{auto_tolerance}
\alias{auto_tolerance}
\title{Determine the optimal error tolerance based on data}
\usage{
auto_tolerance(purity, coverage, fpr = 0.1, epsilon_range = c(0.01, 0.08))
}
\arguments{
\item{purity}{Data purity (putative).}

\item{coverage}{Data coverage.}

\item{fpr}{Desired false positive rate.}

\item{epsilon_range}{Range of values to constrain $epsilon$.}
}
\value{
The $\epsilon$ value estimated from data, constrained to be in
`epsilon_range`, in order to limit the false positive rate to be `fpr`.
}
\description{
Regression has been used to measure the rate of false positives
from simulated tumours with variable coverage and purity. This allows to determine
an optimal valut of $\epsilon$, parameter `purity_error` of function
\code{analyze_peaks}.
}
\examples{
auto_tolerance(.3, 90)
}
