% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_likelihood_functions.R
\name{BAF_LL}
\alias{BAF_LL}
\title{Computes the likelihood of the BAF of a set of binned SNP sitting on a subclonal segment}
\usage{
BAF_LL(baf_obs, n, k1, k2, purity, ccf, g1, g2)
}
\arguments{
\item{baf_obs}{vector of observed BAF for each bin}

\item{n}{number SNPs/bin}

\item{k1}{karytype subclone 1}

\item{k2}{karytype subclone 2}

\item{purity}{}

\item{ccf}{}

\item{g1}{genotype subclone 1}

\item{g2}{genotype subclone 2}
}
\value{
likelihood of the BAF of a set of binned SNP sitting on a subclonal segment
}
\description{
Computes the likelihood of the BAF of a set of binned SNP sitting on a subclonal segment
}
