% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{boostDM_drivers}
\alias{boostDM_drivers}
\title{Driver mutations released with the boost DM method the DriverDBv3 database
\url{https://www.intogen.org/boostdm/}}
\format{
A tibble.
}
\usage{
data(boostDM_drivers)
}
\description{
Driver mutations released with the boost DM method the DriverDBv3 database
\url{https://www.intogen.org/boostdm/}
}
\examples{
data(boostDM_drivers)
boostDM_drivers

}
\references{
Muiños, F., Martínez-Jiménez, F., Pich, O. et al.
In silico saturation mutagenesis of cancer genes. Nature 596, 428–432 (2021).
https://doi.org/10.1038/s41586-021-03771-1
}
\keyword{datasets}
