% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{CCF}
\alias{CCF}
\title{Extract CCF estimates.}
\usage{
CCF(x)
}
\arguments{
\item{x}{A CNAqc object.}
}
\value{
A mutations tibble with columns for mutation multiplicity and CCFs.
}
\description{
This function extracts Cancer Cell Fractions (CCFs) estimates
from a CNAqc object, after they have been computed using
function `compute_CCF`. The estimates are pooled
across the used karyotypes, and extracted from the
`CCF_estimates` field of the input object.
}
\examples{
data("example_PCAWG", package = 'CNAqc')
CCF(example_PCAWG)
}
