% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks_multisample.R
\name{checking_input}
\alias{checking_input}
\title{printing errors, warnings and processing for multi_CNAqc objects and creation}
\usage{
checking_input(cnaqc_objs)
}
\description{
printing errors, warnings and processing for multi_CNAqc objects and creation
}
