% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr_coordinates_GRCh38}
\alias{chr_coordinates_GRCh38}
\title{Coordinates for GRCh38 chromosomes.}
\format{A tibble that represents the coordinates for GRCh38 genome assembly, reporting the chromosome label,
from and to (chromosome range), the length of the chromosome, the position (start and end) of the 
centromers.}
\usage{
data(chr_coordinates_GRCh38)
}
\description{
Coordinates for GRCh38 chromosomes.
}
\examples{
data(chr_coordinates_GRCh38)
chr_coordinates_GRCh38

}
\references{
GRCh38
}
\keyword{datasets}
