% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_likelihood_functions.R
\name{clonal_baf_ll}
\alias{clonal_baf_ll}
\title{Computes the Likelihood of the BAF of a set of SNP falling on a single segment}
\usage{
clonal_baf_ll(baf_obs, n, k1, purity)
}
\arguments{
\item{baf_obs}{vector with the observed BAF for each bin}

\item{n}{number of SNPs / Bin}

\item{k1}{karyotype}

\item{purity}{}
}
\description{
Computes the Likelihood of the BAF of a set of SNP falling on a single segment
}
