% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_likelihood_functions.R
\name{clonal_dr_ll}
\alias{clonal_dr_ll}
\title{Computes likelihood for the depth ratio of a set of binned SNPs on a single segment}
\usage{
clonal_dr_ll(dpr_obs, n, k1, purity, ploidy = 2)
}
\arguments{
\item{dpr_obs}{vector of DR for each SNP}

\item{n}{number of SNPs / bin}

\item{k1}{karyotype}

\item{purity}{}
}
\description{
Computes likelihood for the depth ratio of a set of binned SNPs on a single segment
}
