% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_likelihood_functions.R
\name{clonal_expected_baf}
\alias{clonal_expected_baf}
\title{Computes expected BAF}
\usage{
clonal_expected_baf(k1, purity)
}
\arguments{
\item{k1}{karyotype (ex '1:1', '1:0'...)}

\item{purity}{}
}
\value{
the expected BAF
}
\description{
Computes expected BAF
}
