% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_likelihood_functions.R
\name{clonal_expected_dr}
\alias{clonal_expected_dr}
\title{Computes expected depth ratio}
\usage{
clonal_expected_dr(k1, purity, ploidy = 2)
}
\arguments{
\item{k1}{}

\item{purity}{}
}
\description{
Computes expected depth ratio
}
