% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_patch.R
\name{clonal_test}
\alias{clonal_test}
\title{Clonal test for a segment}
\usage{
clonal_test(SNP_df, SNV_df, purity, ploidy = 2)
}
\arguments{
\item{SNP_df}{data frame of SNPs sitting on the segment}

\item{SNV_df}{dataframe of SNVs sitting on the segment}

\item{purity}{}
}
\value{
a dataframe of the possible karyotype and corresponding expected peaks given a segment, sorted by their likelihood
}
\description{
Clonal test for a segment
}
