% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCF.R
\name{CNA}
\alias{CNA}
\title{Extract CNA information.}
\usage{
CNA(x, type = c("clonal", "subclonal"))
}
\arguments{
\item{x}{A CNAqc object.}

\item{type}{\code{"clonal"} for clonal CNAs, \code{"subclonal"} for subclonal CNAs.}
}
\value{
A tibble with the data.
}
\description{
Getter to obtain copy number calls from an object.
}
\examples{
data("example_PCAWG", package = 'CNAqc')
CNA(example_PCAWG)
}
