% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf.R
\name{CNA_gene}
\alias{CNA_gene}
\title{Extract per-gene copy number status.}
\usage{
CNA_gene(x, genes = NULL)
}
\arguments{
\item{x}{A CNAqc object.}

\item{genes}{Optional, a vector of gene symbols of interest. If `NULL`, all
the human genes are used, according to the genome reference of input `x`.}
}
\value{
A tibble with columns `gene` (gene name), `from`/`to` (gene delimiters),
`Major`/`minor`/`karyotype` as the information for the copy number segment
where the gene sits. Note that if the gene maps to a subclonal segment
this is not returned.
}
\description{
By default, this function maps a list of genes to their copy number status
using clonal CNA segments. The genes used are all the known human genes,
whose coordinates are available inside the CNAqc package and are identified
by common names (e.g., TP53). The function can restrict to a subset of genes
(faster computation) if one passes, via the `genes` parameter, a vector of
gene symbols.
}
\examples{
# Example input data released with the package
data('example_dataset_CNAqc', package = 'CNAqc')
print(example_dataset_CNAqc)

# Note the outputs to screen
x = init(mutations = example_dataset_CNAqc$mutations, cna = example_dataset_CNAqc$cna, purity = example_dataset_CNAqc$purity)

# Get mapping for all the known human genes - takes a bit longer
CNA_gene(x)

# Use known genes set
CNA_gene(x, genes = c("APC", "KRAS", "NRAS", "TP53"))
}
