% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNAqc-package.R
\docType{package}
\name{CNAqc-package}
\alias{CNAqc}
\alias{CNAqc-package}
\title{CNAqc: CNAqc - Copy Number Analysis quality check}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

CNAqc is a package to provide a set of metrics to quality control (QC) allele-specific Copy Number Alteration (CNA) calls, somatic mutations and tumour purity determined from bulk sequencing assays. The package provides statistical measures to quantify the concordance between mutation and CNAs, exploiting the combinatorial relation between allelic imbalance in CNA segments and variant allelic frequencies (VAFs) of somatic mutations. Distinct peak matching procedures for VAFs can be used to QC clonal and subclonal copy numbers. Quantitative metrics and plots for data exploration are also available, allowing for an intuitive assessment of the goodness of somatic calls. QC metrics computed by CNAqc can also be used to re-parametrize a copy number caller and increase the quality of the somatic calls; the Sequenza tool is supported by the CNAqc with an automatica calling-QC pipeline. The tools is general and can be applied downstream of any bioinformatics pipeline. Routines to determine Cancer Cell Franctions (CCFs) for each mutation are also available, along with a model of CCF uncertainty that can be used to identify high-quality CCF calls for downstream analyses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/caravagnalab/CNAqc}
  \item Report bugs at \url{https://github.com/caravagnalab/CNAqc/issues}
}

}
\author{
\strong{Maintainer}: Giulio Caravagna \email{gcaravagn@gmail.com} (\href{https://orcid.org/0000-0003-4240-3265}{ORCID})

Other contributors:
\itemize{
  \item Riccardo Bergamin \email{ric.b.91@gmail.com} [contributor]
  \item Nicola Calonaci \email{nicola.calonaci@gmail.com} [contributor]
  \item Salvatore Milite \email{militesalvatore@gmail.com} [contributor]
  \item Alice Antonello \email{aliceantonello9@gmail.com} [contributor]
}

}
\keyword{internal}
