% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_CCF.R
\name{compute_CCF}
\alias{compute_CCF}
\title{Compute CCF values for the available mutations.}
\usage{
compute_CCF(
  x,
  karyotypes = c("1:0", "1:1", "2:0", "2:1", "2:2"),
  muts_per_karyotype = 25,
  cutoff_QC_PASS = 0.1,
  method = "ENTROPY"
)
}
\arguments{
\item{x}{An object of class \code{cnaqc}, created by the \code{init} function.}

\item{karyotypes}{The karyotypes to use, this package supports only \code{c('2:1', '2:0', '2:2')}.}

\item{muts_per_karyotype}{Minimum number of mutations that are required to be mapped to a karyotype
in order to compute CCF values (default 25).}

\item{cutoff_QC_PASS}{Percentage of mutations that can be not-assigned (\code{NA}) in a karyotype.
If the karyotype has more than \code{cutoff_QC_PASS} percentage of non-assigned mutations then
the overall set of CCF calls is failed for the karyotype.}

\item{method}{Either \code{"ENTROPY"} or \code{"ROUGH"}, to reflect the two different algorithms
to compute CCF. See the package vignette to understand the differences across methods.}
}
\value{
An object of class \code{cnaqc}, with CCF values available for extraction and plotting.
}
\description{
This function provides the implementation of a set of entropy-related routines that can estimate
Cancer Cell Fraction values (CCFs) for the available mutations. The implemented routine is described
in the package vignette `"Computation of Cancer Cell Fractions"` that is available at the URL
\url{https://caravagn.github.io/CNAqc/articles/ccf_computation.html}. This function creates a field
`CCF_estimates` inside the returned object which contains both the estimated CCF values and the
plot of the report of this analysis.
}
\examples{

data('example_dataset_CNAqc')
x = init(example_dataset_CNAqc$snvs, example_dataset_CNAqc$cna, example_dataset_CNAqc$purity)

x = compute_CCF(x, karyotypes = c('1:0', '1:1', '2:1', '2:0', '2:2'))
print(x)

# Extract the values with these other functions
CCF(x)
plot_CCF(x)
}
\seealso{
Getters function \code{CCF} and \code{plot_CCF}.
}
