% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_plot_patch.R
\name{compute_expected_baf_dr_df}
\alias{compute_expected_baf_dr_df}
\title{Title}
\usage{
compute_expected_baf_dr_df(x, subclonal_cna = TRUE, what = "best")
}
\arguments{
\item{x}{cnaqc object}

\item{subclonal_cna}{boolean, whether or not the patch function has been applied}

\item{what}{string, either 'best' or 'subclonal', relevant only if the patch function has been run}
}
\value{
dataframe. Computes the expected BAF and DR for the current CNA calls. If the patch function has 
been run, it will compute the BAF/DR corresponding both to the new and the old calls, otherwise it will compute exclusively 
those corresponding to the input CNAqc calls.
}
\description{
Title
}
