% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_wg_overfragmentation.R
\name{detect_wg_overfragmentation}
\alias{detect_wg_overfragmentation}
\title{Title}
\usage{
detect_wg_overfragmentation(
  x,
  alpha = 0.05,
  binomial_p = 0.2,
  small_segments_cutoff = 1e+06
)
}
\arguments{
\item{alpha}{}
}
\value{

}
\description{
Title
}
