% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_likelihood_functions.R
\name{DR_LL}
\alias{DR_LL}
\title{Computes the likelihood of the DR of a set of binned SNP sitting on a subclonal segment}
\usage{
DR_LL(dpr_obs, n, k1, k2, purity, ccf, ploidy = 2)
}
\arguments{
\item{dpr_obs}{vector of observed DR for each bin}

\item{n}{number of SNPs/bin}

\item{k1}{karytype subclone 1}

\item{k2}{karytype subclone 2}

\item{purity}{}

\item{ccf}{}

\item{ploidy}{average ploidy of the sample, default = 2}
}
\value{
likelihood of the DR of a set of binned SNP sitting on a subclonal segment
}
\description{
Computes the likelihood of the DR of a set of binned SNP sitting on a subclonal segment
}
