% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DriverDBv3}
\alias{DriverDBv3}
\title{Driver mutations released with the DriverDBv3 database
\url{http://driverdb.tms.cmu.edu.tw/}}
\format{
A tibble.
}
\usage{
data(DriverDBv3)
}
\description{
Driver mutations released with the DriverDBv3 database
\url{http://driverdb.tms.cmu.edu.tw/}
}
\examples{
data(DriverDBv3)
DriverDBv3

}
\references{
??
}
\keyword{datasets}
