% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dataset_CNAqc}
\alias{example_dataset_CNAqc}
\title{Example CNAqc dataset.}
\format{
A list of SNVs, allele-specific copy number alterations (CNAs) and purity value that
can be used with CNAqc. This tumour does not contain subclonal CNAs.
}
\usage{
data(example_dataset_CNAqc)
}
\description{
Example CNAqc dataset.
}
\examples{
data(example_dataset_CNAqc)
example_dataset_CNAqc

}
\keyword{datasets}
