% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_PCAWG}
\alias{example_PCAWG}
\title{Example PCAWG tumour}
\format{
A CNAqc object created for a PCAWG sample, with subclonal CNAs
called by Battenberg.
}
\usage{
data(example_PCAWG)
}
\description{
Example PCAWG tumour
}
\examples{
data(example_PCAWG)
example_PCAWG

}
\references{
The ICGC/TCGA Pan-Cancer Analysis of Whole Genomes Consortium.
Pan-cancer analysis of whole genomes. Nature 578, 82–93 (2020).
 https://doi.org/10.1038/s41586-020-1969-6
}
\keyword{datasets}
