% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_likelihood_functions.R
\name{expected_baf}
\alias{expected_baf}
\title{Computes expected BAF in the case of two subclones}
\usage{
expected_baf(k1, k2, purity, ccf, g1, g2)
}
\arguments{
\item{k1}{karyorype subclone 1}

\item{k2}{karyorype subclone 2}

\item{purity}{}

\item{ccf}{}

\item{g1}{genotype subclone 1 (the mathernal and paternal alleles are named A and B, plus a number indicating when they were originated : A1B1 = 1:1 )}

\item{g2}{genotype subclone 2}
}
\value{
expected BAF
}
\description{
Computes expected BAF in the case of two subclones
}
