% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_likelihood_functions.R
\name{expected_dr}
\alias{expected_dr}
\title{Computes expected DR in the case of two subclones}
\usage{
expected_dr(k1, k2, purity, ccf, ploidy = 2)
}
\arguments{
\item{k1}{karytype subclone 1}

\item{k2}{karytype subclone 2}

\item{purity}{}

\item{ccf}{}

\item{ploidy}{average ploidy of the sample, default = 2}
}
\value{
expected DR in the case of two subclones
}
\description{
Computes expected DR in the case of two subclones
}
