% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_processing_functions.R
\name{extract_sequenza_baf_dr}
\alias{extract_sequenza_baf_dr}
\title{Get SNPs dataframe from Sequenza output}
\usage{
extract_sequenza_baf_dr(x)
}
\arguments{
\item{x}{the result of the following operation:
x = load('[sample_name]_sequenza_extract.RData')
x = get(x)
where the _sequenza_extract.RData file can be found among the outputs of Sequenza}
}
\value{
a list with average BAF and DR either for each bin ($binned) or for each segment ($segment)
}
\description{
Get SNPs dataframe from Sequenza output
}
