% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fpr_test}
\alias{fpr_test}
\title{#' Example PCAWG tumour
#'
#' @docType data
#'
#' @usage data(example_PCAWG)
#'
#' @format A CNAqc object created for a PCAWG sample, with subclonal CNAs
#' called by Battenberg.
#'
#' @keywords datasets
#'
#' @references The ICGC/TCGA Pan-Cancer Analysis of Whole Genomes Consortium.
#' Pan-cancer analysis of whole genomes. Nature 578, 82–93 (2020).
#'  https://doi.org/10.1038/s41586-020-1969-6
#' @examples
#' data(example_PCAWG)
#' example_PCAWG
#'
"example_PCAWG"
Data (simulation performance) from the trainig set to auto-tune epsilon.}
\format{
A tibble.
}
\usage{
data(fpr_test)
}
\description{
#' Example PCAWG tumour
#'
#' @docType data
#'
#' @usage data(example_PCAWG)
#'
#' @format A CNAqc object created for a PCAWG sample, with subclonal CNAs
#' called by Battenberg.
#'
#' @keywords datasets
#'
#' @references The ICGC/TCGA Pan-Cancer Analysis of Whole Genomes Consortium.
#' Pan-cancer analysis of whole genomes. Nature 578, 82–93 (2020).
#'  https://doi.org/10.1038/s41586-020-1969-6
#' @examples
#' data(example_PCAWG)
#' example_PCAWG
#'
"example_PCAWG"
Data (simulation performance) from the trainig set to auto-tune epsilon.
}
\examples{
data(fpr_test)
fpr_test

}
\keyword{datasets}
