% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_coordinates_GRCh38}
\alias{gene_coordinates_GRCh38}
\title{Coordinates for GRCh38 genes}
\format{
A tibble that represents the coordinates for genes in the GRCh38 genome assembly,
reporting the gene symbol, from and to (chromosome range).
}
\usage{
data(gene_coordinates_GRCh38)
}
\description{
Coordinates for GRCh38 genes
}
\examples{
data(gene_coordinates_GRCh38)
gene_coordinates_GRCh38

}
\references{
GRCh38
}
\keyword{datasets}
