% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_coordinates_hg19}
\alias{gene_coordinates_hg19}
\title{Coordinates for hg19 genes}
\format{
A tibble that represents the coordinates for genes in the hg19 genome assembly,
reporting the gene symbol, from and to (chromosome range).
}
\usage{
data(gene_coordinates_hg19)
}
\description{
Coordinates for hg19 genes
}
\examples{
data(gene_coordinates_hg19)
gene_coordinates_hg19

}
\references{
hg19
}
\keyword{datasets}
