% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_simulations_functions.R
\name{generate_clonal_snvs}
\alias{generate_clonal_snvs}
\title{allele'= allele,}
\usage{
generate_clonal_snvs(
  cna,
  seg_id,
  karyotype,
  chromosome,
  mu = 1e-08,
  w = 2,
  t_subclone_mrca = 5,
  coverage = 100,
  purity = 1,
  t_cna
)
}
\description{
allele'= allele,
}
