% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_patch.R
\name{get_clonal_peaks}
\alias{get_clonal_peaks}
\title{Get clonal peaks}
\usage{
get_clonal_peaks(k, purity)
}
\arguments{
\item{k}{simple clonal karyotype}

\item{purity}{}
}
\value{
expected peaks
}
\description{
Get clonal peaks
}
