% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drivers.R
\name{get_drivers}
\alias{get_drivers}
\title{Extract drivers data.}
\usage{
get_drivers(x, chromosomes = paste0("chr", c(1:22, "X", "Y")), which = "VAF")
}
\arguments{
\item{x}{A CNAqc object.}

\item{chromosomes}{Which chromosomes to subset.}

\item{which}{A keyword for \code{"VAF"} or \code{"CCF"}.}
}
\value{
A tibble
}
\description{
Returns a tibble with the drivers annotated on certain chromosomes,
extracting VAF or CCF values (if computed). If drivers are not
annotated \code{NULL} is returned.
}
\examples{
data('example_dataset_CNAqc', package = 'CNAqc')
x = init(mutations = example_dataset_CNAqc$mutations, cna = example_dataset_CNAqc$cna, purity = example_dataset_CNAqc$purity)

get_drivers(x)
}
