% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_m_cnaqc.R
\name{get_mCNAqc_stats}
\alias{get_mCNAqc_stats}
\title{Get mCNAqc statistics}
\usage{
get_mCNAqc_stats(m_cnaqc_obj)
}
\arguments{
\item{m_cnaqc_obj}{}
}
\value{
a tibble
}
\description{
Get mCNAqc statistics
}
\examples{
\dontrun{get_mCNAqc_stats(x)}

}
