% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_patch.R
\name{get_models}
\alias{get_models}
\title{Get model}
\usage{
get_models(karyotype_combination = NULL, purity = NULL, ccf = NULL)
}
\arguments{
\item{karyotype_combination}{combination of simple karyotypes of the two subclones}

\item{purity}{purity to filter result}

\item{ccf}{}
}
\value{
dataframe with the possible evolutionary models and corresponding expected peaks
}
\description{
Get model
}
