% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_PASS_percentage}
\alias{get_PASS_percentage}
\title{Returns percentage of passed segments}
\usage{
get_PASS_percentage(x)
}
\arguments{
\item{x}{A CNAqc object.}
}
\value{
A tibble with the data.
}
\description{
Getter to the percentage of genome with pass value
}
\examples{
data("example_PCAWG", package = 'CNAqc')
get_PASS_percentage(example_PCAWG)
}
