% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_m_cnaqc.R
\name{get_sample}
\alias{get_sample}
\title{get desidered samples}
\usage{
get_sample(m_cnaqc_obj, sample, which_obj)
}
\arguments{
\item{m_cnaqc_obj}{mCNAqc object.}

\item{sample}{list of one or more samples, must correspond to the names of each element in the mCNAqc object}

\item{which_obj}{one of \code{shared} or \code{original}.}
}
\value{
a list of CNAqc objects corresponding to the desidered samples.
}
\description{
get desidered samples
}
\examples{
\dontrun{get_sample(x, sample = get_sample_names(x), which_obj = "shared") }

}
