% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_m_cnaqc.R
\name{get_sample_name}
\alias{get_sample_name}
\title{get sample names from CNAqc and mCNAqc objects}
\usage{
get_sample_name(x)
}
\arguments{
\item{x}{a CNAqc or a mCNAqc object.}
}
\value{
a vector with sample names of CNAqc objects.
}
\description{
get sample names from CNAqc and mCNAqc objects
}
\examples{
\dontrun{get_sample_name(x)}

}
