% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_processing_functions.R
\name{get_segment_type}
\alias{get_segment_type}
\title{Returns the segment type of a segment based on CCF and karyotype}
\usage{
get_segment_type(x, seg_id)
}
\arguments{
\item{x}{a CNAqc object}

\item{seg_id}{the id of the segment}
}
\value{
the segment type according to CCF and karyotype: 'simple clonal', 'simple subclonal','complex clonal','complex subclonal'
}
\description{
Returns the segment type of a segment based on CCF and karyotype
}
