% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_segment.R
\name{inspect_segment}
\alias{inspect_segment}
\title{Plot VAFs across chromosomes.}
\usage{
inspect_segment(
  x,
  chrs = c(paste0("chr", 1:22), "chrX", "chrY"),
  n = 200,
  l = 1000
)
}
\arguments{
\item{x}{A CNAqc object.}

\item{chrs}{The chromosome ids to use (e.g., "chr2"). All are used by default.}

\item{n}{Disregard chromosomes with less than `n` mutations mapped.}

\item{l}{Disregard chromosomes that span less than `l` nucleotide.}
}
\value{
A `ggplot2` object.
}
\description{
Plot VAFs across chromosomes for mutations mapping to
clonal simple CNAs. FIlters can be used to subset the data. Every segment has
a colour, and  maximum 74 segements can be plotted. The plots are
split by chromosome and karyotype.
}
\examples{
data('example_dataset_CNAqc', package = 'CNAqc')
x = init(mutations = example_dataset_CNAqc$mutations, cna = example_dataset_CNAqc$cna, purity = example_dataset_CNAqc$purity)

# Deafault segments -- all chromsomes, at least 1KB, at least 200 mapped mutations.
inspect_segment(x)

# Same as above, but only for chromosome 2
inspect_segment(x, chrs = 'chr2', n = 50)
}
