% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intogen_drivers}
\alias{intogen_drivers}
\title{List of Intogen driver genes per tumour type.}
\format{
A tibble.
}
\usage{
data(intogen_drivers)
}
\description{
List of Intogen driver genes per tumour type.
}
\examples{
data(intogen_drivers)
intogen_drivers

}
\references{

}
\keyword{datasets}
