% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_ccf_purity}
\alias{model_ccf_purity}
\title{Expected VAF for different evolutionary models}
\format{
a nested list of dataframes reporting all expected VAF peaks computed for different 
combinations of karyotipes,
}
\usage{
data(model_ccf_purity)
}
\description{
Expected VAF for different evolutionary models
}
\examples{
data(model_ccf_purity)
model_ccf_purity

}
\keyword{datasets}
