% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{Mutations}
\alias{Mutations}
\title{Extract mutations.}
\usage{
Mutations(x, cna = c("clonal", "subclonal"), type = c("SNV", "indel"))
}
\arguments{
\item{x}{A CNAqc object.}

\item{cna}{\code{"clonal"} for clonal CNAs, \code{"subclonal"} for subclonal CNAs.}

\item{type}{\code{"SNV"} for single-nucleotide variants, \code{"indel"} for insertion-deletions.}
}
\value{
A tibble with the data.
}
\description{
Getter to obtain mutation calls from an object.
}
\examples{
data("example_PCAWG", package = 'CNAqc')
Mutations(example_PCAWG)
}
