% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_patch.R
\name{patch}
\alias{patch}
\title{Given a vector of segments to test, finds the best copy number solution based on VAF, BAF and DR}
\usage{
patch(
  x,
  segments = NULL,
  top_n = 5,
  all_solutions = TRUE,
  preselect = FALSE,
  baf_coef = 30,
  dr_coef = 10
)
}
\arguments{
\item{x}{a cnaqc object}

\item{segments}{optional, a user-defined vector of segment ids}

\item{top_n}{optional, alternative to all_solutions, integer declaring how many solutions are to be saved}

\item{all_solutions}{optional, alternative to top_n, TRUE/FALSE whether to save all solutions}

\item{preselect}{optional, whether to pre-select the segment to be tested based on BAF and DR}

\item{baf_coef}{otional, indicates the weight of the measurement in choosing the best solution (the higher the weigh, the lower the relevance)}

\item{dr_coef}{}
}
\value{
the cnaqc object with three new fields : 
             - patch_best_solution : the best solutions for each tested segment
             - patch_clonal_solution : top n / all clonal solutions for the tested segments
             - patch_sunclonal_solution : top n / all subclonal solutions for the tested segments
}
\description{
Given a vector of segments to test, finds the best copy number solution based on VAF, BAF and DR
}
