% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_plot_patch.R
\name{patch_plot}
\alias{patch_plot}
\title{Plot summarising the results of patch for a single segment (best solution, all subclonal solutions, all clonal solutions)}
\usage{
patch_plot(x, seg_id)
}
\arguments{
\item{x}{a cnaqc object}

\item{seg_id}{the segment id of the segment of interest}
}
\description{
Plot summarising the results of patch for a single segment (best solution, all subclonal solutions, all clonal solutions)
}
