% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_plot_patch.R
\name{patch_plot_segments}
\alias{patch_plot_segments}
\title{Title}
\usage{
patch_plot_segments(x, subclonal_cna = TRUE, what = "best", s = 1, a = 0.6)
}
\arguments{
\item{x}{cnaqc object}

\item{subclonal_cna}{boolean, whether or not the patch function has been applied}

\item{what}{string, either 'best' or 'subclonal', relevant only if the patch function has been run}

\item{s}{size of the expected BAF/DR segments}

\item{a}{alpha of the expected BAF/DR segments}
}
\value{
plot. If the patch function has been run, it will show how
the expected BAF and DR of the newly called CNAs match the one observed in data. Otherwise, it
will show how the expected BAF and DR of the original call match the observed data
}
\description{
Title
}
