% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cnaqc.R
\name{plot.cnaqc}
\alias{plot.cnaqc}
\title{Plot for class \code{'cnaqc'}.}
\usage{
\method{plot}{cnaqc}(x, ...)
}
\arguments{
\item{x}{An obj of class \code{'cnaqc'}.}

\item{...}{Default S3 method parameter.}
}
\value{
Nothing.
}
\description{
The default plot is the CNA segments in wide format
}
\examples{
data('example_PCAWG', package = 'CNAqc')

plot(example_PCAWG)
}
