% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.m_cnaqc.R
\name{plot.m_cnaqc}
\alias{plot.m_cnaqc}
\title{Plot for class \code{'m_cnaqc'}.}
\usage{
\method{plot}{m_cnaqc}(x, ...)
}
\arguments{
\item{x}{An obj of class \code{'m_cnaqc'}.}

\item{...}{Default S3 method parameter.}
}
\value{
Nothing.
}
\description{
The default plot is the CNA segments in wide format
}
\examples{
\dontrun{

}

}
