% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CCF.R
\name{plot_CCF}
\alias{plot_CCF}
\title{Plot the CCF estimates in the data.}
\usage{
plot_CCF(x)
}
\arguments{
\item{x}{An object of class \code{cnaqc}, where CCF have been computed using
function `compute_CCF`.}
}
\value{
A `ggpubr` figure of the CCF estiamtes.
}
\description{
This function extracts the plots of the Cancer Cell Fractions (CCFs) estimates
from a `CNAqc` object, after they have been computed using
function `compute_CCF`. The plots - one per computed karyotype - are assembled
into a `ggpubr` figure.
}
\examples{
data('example_dataset_CNAqc')
x = init(example_dataset_CNAqc$snvs, example_dataset_CNAqc$cna, example_dataset_CNAqc$purity)

x = compute_CCF(x)
plot_CCF(x)
}
