% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_plot_patch.R
\name{plot_dr_single_segment}
\alias{plot_dr_single_segment}
\title{Plot the DR of bins sitting on a selected segment}
\usage{
plot_dr_single_segment(SNP_df)
}
\arguments{
\item{SNP_df}{dataframe with the binned SNPs sitting on the segment}
}
\description{
Plot the DR of bins sitting on a selected segment
}
