% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome_wide_mutations.R
\name{plot_gw_vaf}
\alias{plot_gw_vaf}
\title{Plot genome-wide VAFs.}
\usage{
plot_gw_vaf(x, N = 5000, chromosomes = paste0("chr", c(1:22, "X", "Y")))
}
\arguments{
\item{x}{A CNAqc object.}

\item{N}{Mutations to use, randomly sampled.}

\item{chromosomes}{The chromosomes to use for this plot.}
}
\value{
A \code{ggplot2} plot.
}
\description{
@description Plot a genome-wide scatter plot of mutation VAFs.
The function can randomly subset the points to show.
}
\examples{
data('example_dataset_CNAqc', package = 'CNAqc')
x = init(mutations = example_dataset_CNAqc$mutations, cna = example_dataset_CNAqc$cna, purity = example_dataset_CNAqc$purity)

plot_gw_vaf(x)

# Downsampling examples
plot_gw_vaf(x, N = 100)
plot_gw_vaf(x, N = 1000)
plot_gw_vaf(x, N = 10000)
}
