% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_karyotypes.R
\name{plot_karyotypes}
\alias{plot_karyotypes}
\title{Title}
\usage{
plot_karyotypes(x, type = "percentage", chromosomes = paste0("chr",
  c(1:22, "X", "Y")))
}
\arguments{
\item{chromosomes}{}
}
\description{
Title
}
