% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_plot_patch.R
\name{plot_patch_all_solutions_clonal}
\alias{plot_patch_all_solutions_clonal}
\title{Plot summarising the clonal solutions for a single segment found by patch}
\usage{
plot_patch_all_solutions_clonal(x, seg_id)
}
\arguments{
\item{x}{a cnaqc object}

\item{seg_id}{}
}
\description{
Plot summarising the clonal solutions for a single segment found by patch
}
