% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_plot_patch.R
\name{plot_patch_all_solutions_subclonal}
\alias{plot_patch_all_solutions_subclonal}
\title{Plot summarising all subclonal solutions for a single segment found by patch in terms of loglikelihood and CCF, dividing them by karyotype combination and model type}
\usage{
plot_patch_all_solutions_subclonal(x, seg_id)
}
\arguments{
\item{x}{a cnaqc object}

\item{seg_id}{the segment of interest}
}
\description{
Plot summarising all subclonal solutions for a single segment found by patch in terms of loglikelihood and CCF, dividing them by karyotype combination and model type
}
