% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_plot_patch.R
\name{plot_patch_best_solution}
\alias{plot_patch_best_solution}
\title{Plot comparing the original solution with the best one found by function patch (BAF, DR, VAF)}
\usage{
plot_patch_best_solution(x, seg_id)
}
\arguments{
\item{x}{cnaqc object}

\item{seg_id}{id of the segment of interest}
}
\description{
Plot comparing the original solution with the best one found by function patch (BAF, DR, VAF)
}
