% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_sbs_contexts.R
\name{plot_SBS}
\alias{plot_SBS}
\title{Plost SBS counts}
\usage{
plot_SBS(x, highlight = 0.05)
}
\arguments{
\item{x}{A CNAqc object.}

\item{highlight}{Minimum frequency to color in red a bar of the barplot (default 5%).}
}
\value{
A ggplot object.
}
\description{
Canonical plot of SBS count from a sample, obtained
upon mapping substitutions to contexts via function \code{SBS}.

The function allows to highlight in red a substitution that occurs with
a frequency above a desired cutoff (default 5%).
}
\examples{
data('example_dataset_CNAqc')
x = init(mutations = example_dataset_CNAqc$mutations, cna =example_dataset_CNAqc$cna, purity = example_dataset_CNAqc$purity)

x = SBS(x)

# SBS plot
plot_SBS(x)
}
