% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_segment_size_distribution.R
\name{plot_segment_size_distribution}
\alias{plot_segment_size_distribution}
\title{Title}
\usage{
plot_segment_size_distribution(x, annotate_cut = 1e+07)
}
\arguments{
\item{annotate_cut}{}
}
\description{
Title
}
