% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mcnaqc_cna.R
\name{plot_segments_multisample}
\alias{plot_segments_multisample}
\title{Plot multisample cna data}
\usage{
plot_segments_multisample(
  x,
  which,
  chromosomes = paste0("chr", c(1:22, "X", "Y"))
)
}
\arguments{
\item{x}{A m_cnaqc object}

\item{which}{either "original" or "shared"}

\item{chromosomes}{}
}
\value{
a plot
}
\description{
Plot multisample cna data
}
\examples{
\dontrun{plot_segments_multisample(x, which = "shared")}
}
