% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smoothing.R
\name{plot_smoothing}
\alias{plot_smoothing}
\title{Plot smoothed and non-smoohted segments.}
\usage{
plot_smoothing(x)
}
\arguments{
\item{x}{A CNAq object.}
}
\value{
A \code{ggpubr} figure..
}
\description{
Upon using the `smooth_segments` function, this function plots a multipanel figure
with the segments profile before and after smoothing.
}
\examples{
data('example_dataset_CNAqc', package = 'CNAqc')
x = init(mutations = example_dataset_CNAqc$mutations, cna = example_dataset_CNAqc$cna, purity = example_dataset_CNAqc$purity)

# Smooth
x = smooth_segments(x)

# View differences
plot_smoothing(x)
}
\seealso{
smooth_segments
}
