% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_processing_functions.R
\name{plot_snps}
\alias{plot_snps}
\title{Plot BAF and Depth Ratio of binned SNPs}
\usage{
plot_snps(x, what = "BAF", s = 1)
}
\arguments{
\item{x}{cnaqc object}

\item{what}{either BAF of DR}
}
\value{
the plot of BAF or DR of binned SNPs along the genome
}
\description{
Plot BAF and Depth Ratio of binned SNPs
}
