% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_plot_patch.R
\name{plot_vaf_single_segment}
\alias{plot_vaf_single_segment}
\title{Plot the VAF of SNVs sitting on a single segment}
\usage{
plot_vaf_single_segment(SNV_df)
}
\arguments{
\item{SNV_df}{dataframe of SNVs sitting on the segment}
}
\description{
Plot the VAF of SNVs sitting on a single segment
}
