% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_patch.R
\name{pre_select_segments}
\alias{pre_select_segments}
\title{Pre-filters the segments with suspect copy number calls based on DR and BAF}
\usage{
pre_select_segments(x)
}
\arguments{
\item{x}{a cnaqc object}
}
\value{
a vector with the pre-selected segment ids
}
\description{
Pre-filters the segments with suspect copy number calls based on DR and BAF
}
