% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cnaqc.R
\name{print.cnaqc}
\alias{print.cnaqc}
\title{Print for class \code{'cnaqc'}.}
\usage{
\method{print}{cnaqc}(x, ...)
}
\arguments{
\item{x}{An obj of class \code{'cnaqc'}.}

\item{...}{Default S3 method parameter.}
}
\value{
Nothing.
}
\description{
Print for class \code{'cnaqc'}.
}
\examples{
data('example_PCAWG', package = 'CNAqc')

print(example_PCAWG)
}
