% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.m_cnaqc.R
\name{print.m_cnaqc}
\alias{print.m_cnaqc}
\title{Print for class \code{'m_cnaqc'}.}
\usage{
\method{print}{m_cnaqc}(x, ...)
}
\arguments{
\item{x}{An obj of class \code{'m_cnaqc'}.}

\item{...}{Default S3 method parameter.}
}
\value{
Nothing.
}
\description{
Print for class \code{'m_cnaqc'}.
}
\examples{

### search for an example to add here and include in the package

\dontrun{
data('example_multisample', package = 'CNAqc')

print(example_multisample)
}

}
