% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_processing_functions.R
\name{process_snps}
\alias{process_snps}
\title{Maps SNPs bins onto segments}
\usage{
process_snps(snps, cna, subclonal_cna = NULL)
}
\arguments{
\item{snps}{A dataframe of single nucleotide polymorphism mean BAF and DR over fixed genomic windows (bins) with the following fields:

* `chr` chromosome name, e.g., \code{"chr3"}, \code{"chr8"}, \code{"chrX"}, ...
* `from` where the bin start, an integer number
* `to` where the bin ends, an integer number
* `BAF` average B-allele frequency of the bin, a real number in [0,1]
* `N.BAF` number if SNPs falling into the bin
* `DR` average depth ration of the bin, a real positive number}

\item{cna}{A dataframe of allele-specific copy number with the following fields:

* `chr` chromosome name, e.g., \code{"chr3"}, \code{"chr8"}, \code{"chrX"}, ...
* `from` where the segment start, an integer number
* `to` where the segment ends, an integer number
* `Major` for the number of copies of the major allele (or A-allele), an integer number
* `minor` for the number of copies of the major allele (or B-allele), an integer number
* `CCF` an optional cancer cell fraction (CCF) column distinguishing clonal and subclonal segments, a real number in [0,1]
* `Major_2` optional for the number of copies of the major allele (or A-allele) in the second clone if present, an integer number
* `minor_2` optional for the number of copies of the major allele (or B-allele) in the second clone  if present, an integer number}

\item{subclonal_cna}{same as cna, but for cna segments with ccf< 1}
}
\value{
the snps object with an extra column, `segment_id` reporting the id of the segment they map on
}
\description{
Maps SNPs bins onto segments
}
