% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_sbs_contexts.R
\name{SBS}
\alias{SBS}
\title{Augment SBS data for mutational signatures deconvolution}
\usage{
SBS(x)
}
\arguments{
\item{x}{A CNAqc object.}
}
\value{
A CNAqc object with required SBS data in the inner field `SBS`.
}
\description{
Augment a CNAqc object with Single Base Substitution (SBS) data for
mutational signature deconvolution.

This creates set of objects that are stored inside field `SBS` of the input object.
These are:

* `SNVs`, a tibble for SNVs used for SBS data preparation;
* `GRanges`, a GRanges object for the above tibble;
* `counts`, a trinucleotide count matrix for SBS deconvolution.

In particular, `counts` is the canonnical format to run SBS deconvolution
in a variety of tools.
}
\examples{
data('example_dataset_CNAqc')
x = init(mutations = example_dataset_CNAqc$mutations, cna =example_dataset_CNAqc$cna, purity = example_dataset_CNAqc$purity)

x = SBS(x)

# All SBS data
print(x$SBS)
}
