% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters_mcnaqc.R
\name{Shared_segments_mutations}
\alias{Shared_segments_mutations}
\title{Extract mutations in common segments from multi_CNAqc object.}
\usage{
Shared_segments_mutations(
  m_cnaqc_obj,
  sample = "all",
  cna = c("clonal", "subclonal"),
  type = c("SNV", "indel")
)
}
\arguments{
\item{m_cnaqc_obj}{multi_CNAqc object (must have previously run \code{"multisample_init"})}

\item{sample}{desired samples for which you want to extract the mutations. Default \code{"all"}, extracts mutations from all samples
in the multi_CNAqc object.}

\item{cna}{\code{"clonal"} for clonal CNAs, \code{"subclonal"} for subclonal CNAs.}

\item{type}{\code{"SNV"} for single-nucleotide variants, \code{"indel"} for insertion-deletions.}
}
\value{
a named list of tibbles with all the mutations.
}
\description{
This function extracts the all the mutations belonging to common segments in CNAqc format of 
desired samples from a multi_CNAqc object.
}
\examples{
 
shared_segments_mut = Shared_segments_mutations(example_multisample, 
                                sample = "all", 
                                cna = c("clonal", "subclonal"), 
                                type = c("SNV", "indel"))
print(shared_segments_mut)
}
