% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_simulations_functions.R
\name{simulate_subclonal_snvs_df}
\alias{simulate_subclonal_snvs_df}
\title{Major'= sub_cnas$Major_2[i],'minor'=sub_cnas$minor_2[i],}
\usage{
simulate_subclonal_snvs_df(
  sub_cnas,
  mu = 1e-08,
  w = 2,
  t_f = 10,
  coverage = 100,
  purity = 1
)
}
\description{
Major'= sub_cnas$Major_2[i],'minor'=sub_cnas$minor_2[i],
}
