% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_segments.R
\name{smooth_segments}
\alias{smooth_segments}
\title{Smooth simple clonal CNAs.}
\usage{
smooth_segments(x, maximum_distance = 1e+06)
}
\arguments{
\item{x}{A CNAqc object.}

\item{maximum_distance}{The maximum number of nucleotides to  join two equivalent
segments.}
}
\value{
A CNAqc object with smoothed segments, which retains also the pre-smooting
dataset.
}
\description{
This function joins simple clonal CNAs that have the same Major and minor alleles,
and that are zeparate by at most a certain number
of nucleotides. The pre-smoothing copy number segments are
retained in the output computation.
}
\examples{
data('example_dataset_CNAqc', package = 'CNAqc')
x = init(mutations = example_dataset_CNAqc$mutations, cna = example_dataset_CNAqc$cna, purity = example_dataset_CNAqc$purity)

# Before smoothing
print(x)
x = smooth_segments(x)

# After smoothing
plot_smoothing(x)
}
