% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{split_by_chromosome}
\alias{split_by_chromosome}
\title{Split a dataset by chromosome.}
\usage{
split_by_chromosome(x, chromosomes = paste0("chr", c(1:22, "X", "Y")))
}
\arguments{
\item{x}{An object created by CNAqc.}

\item{chromosomes}{A list of chromosomes to retain.}
}
\value{
A named list of CNAqc objects.
}
\description{
Split a CNAqc object by chromosome, returning a named list
with data splits by chromosomes. In this way it is easy to run QC steps
per chromosome.
}
\examples{

data("example_PCAWG", package = 'CNAqc')

# Split of tje PCAWG object, for instance
split_by_chromosome(example_PCAWG)
}
