% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_patch.R
\name{sub_clonal_test}
\alias{sub_clonal_test}
\title{Subclonal test for a segment}
\usage{
sub_clonal_test(SNP_df, SNV_df, purity, ploidy = 2, penalty = 0)
}
\arguments{
\item{SNP_df}{data frame of SNPs sitting on the segment}

\item{SNV_df}{dataframe of SNVs sitting on the segment}

\item{purity}{}
}
\value{
a dataframe of the possible karyotypes and CCF combinations and corresponding expected peaks given a segment, sorted by their likelihood
}
\description{
Subclonal test for a segment
}
