% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subset_by_minimum_CCF}
\alias{subset_by_minimum_CCF}
\title{Subset mutations with minimum CCF values.}
\usage{
subset_by_minimum_CCF(x, min_target_CCF = 0.1)
}
\arguments{
\item{x}{A CNAqc object.}

\item{min_target_CCF}{The minimum CCF do be enforced.}
}
\value{
A new CNAqc object with filtered mutations.
}
\description{
This function allows to subset a CNAqc object and retain only mutations that
map to clonal CNAs and have a mininum CCF value. Note that subclonal CNAs are
lost upon application of this function.
}
\examples{

x = init(mutations = example_dataset_CNAqc$mutations, cna = example_dataset_CNAqc$cna, purity = example_dataset_CNAqc$purity)
print(x)

plot_data_histogram(x)

plot_data_histogram(subset_by_minimum_CCF(x))
}
