% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subset_by_segment_karyotype}
\alias{subset_by_segment_karyotype}
\title{Subset by clonal segments.}
\usage{
subset_by_segment_karyotype(x, karyotypes)
}
\arguments{
\item{x}{A new CNAqc object with subset data.}

\item{karyotypes}{A list of karyotype ids in \code{"Major:minor"} notation
(e.g., \code{"1:1", "2,1", ...}) that will be retained.}
}
\description{
Retains only a subset of clonal segments, e.g., all `2:1` segments.
}
\examples{
data('example_dataset_CNAqc', package = 'CNAqc')
x = init(mutations = example_dataset_CNAqc$mutations, cna = example_dataset_CNAqc$cna, purity = example_dataset_CNAqc$purity)

subset_by_segment_karyotype(x, '2:2')
subset_by_segment_karyotype(x, '2:1')
}
