% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subset_by_segment_minmutations}
\alias{subset_by_segment_minmutations}
\title{Retain clonal segments with minimum number of mutations.}
\usage{
subset_by_segment_minmutations(x, n = 50)
}
\arguments{
\item{x}{A CNAqc object.}

\item{n}{The minimum number of mutations per segment.}
}
\value{
A CNAqc object that should have less segments.
}
\description{
It retains only the set of clonal segments that have mapped
at least `n` mutations mapped.
}
\examples{
data('example_dataset_CNAqc', package = 'CNAqc')
x = init(mutations = example_dataset_CNAqc$mutations, cna = example_dataset_CNAqc$cna, purity = example_dataset_CNAqc$purity)

subset_by_segment_minmutations(x, 0)
subset_by_segment_minmutations(x, 10)
}
