% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subset_by_segment_totalcn}
\alias{subset_by_segment_totalcn}
\title{Subset clonal simple segments total copy number.}
\usage{
subset_by_segment_totalcn(x, totalcn)
}
\arguments{
\item{x}{A CNAqc object.}

\item{totalcn}{The total (Major + minor) copy numbers to filter.}
}
\value{
x A new CNAqc object with subset data.
}
\description{
Retains only a subset of clonal simple segments, e.g., all segments with
total ploidy 2.
}
\examples{
data('example_dataset_CNAqc', package = 'CNAqc')
x = init(mutations = example_dataset_CNAqc$mutations, cna = example_dataset_CNAqc$cna, purity = example_dataset_CNAqc$purity)

subset_by_segment_totalcn(x, 2)
}
