% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subset_snvs}
\alias{subset_snvs}
\title{Subset only SNVs.}
\usage{
subset_snvs(
  x,
  ref_nucleotides = c("A", "C", "T", "G"),
  alt_nucleotides = c("A", "C", "T", "G")
)
}
\arguments{
\item{x}{A CNAqc object.}

\item{ref_nucleotides}{What reference alleles to use, default \code{c("A", "C", "T", "G")}.}

\item{alt_nucleotides}{What alternative alleles to use, default \code{c("A", "C", "T", "G")}.}
}
\value{
A new CNAqc dataset created with \code{init}.
}
\description{
This function removes all non-SNVs mutations, re-creating a new CNAqc
object with just SNVs. All analyses are lost.
}
\examples{
data('example_dataset_CNAqc', package = 'CNAqc')
x = init(mutations = example_dataset_CNAqc$mutations, cna = example_dataset_CNAqc$cna, purity = example_dataset_CNAqc$purity)

subset_snvs(x)
}
