% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_patch.R
\name{test_model}
\alias{test_model}
\title{Test models for a single segment}
\usage{
test_model(
  x,
  seg_id,
  top_n = 5,
  all_solutions = FALSE,
  baf_coef = 10,
  dr_coef = 5
)
}
\arguments{
\item{x}{cnaqc object}

\item{seg_id}{cnaqc object}

\item{top_n}{how many results to output}

\item{purity}{}
}
\description{
Test models for a single segment
}
