% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_patch.R
\name{test_quality_segment}
\alias{test_quality_segment}
\title{Test whether the expected BAF and DR match the observed ones, i.e. if it is advisable to patch the calls of the segment}
\usage{
test_quality_segment(x, seg_id, var = 0.005, significance = 0.05, var_dr = 10)
}
\arguments{
\item{x}{a cnaqc object}

\item{seg_id}{the id of the segment to test}

\item{var}{expected variance for the BAF distribution}

\item{significance}{significance level for the Kolmogorov-Smirnov test}

\item{var_dr}{expected DR rate parameter}
}
\value{
boolean TRUE or FAlSE, wheter the segment had significantly different BAF of DR distributions
}
\description{
Test whether the expected BAF and DR match the observed ones, i.e. if it is advisable to patch the calls of the segment
}
