% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subclonal_cna_likelihood_functions.R
\name{VAF_LL}
\alias{VAF_LL}
\title{Computes the log-likelihood for a set of SNVs falling on a single segment}
\usage{
VAF_LL(NV, DP, peaks)
}
\arguments{
\item{NV}{vector of number of reads in the region carrying the variants (NV)}

\item{DP}{vector of number of reads in the region (DP)}

\item{peaks}{vector of the VAF of the expected peaks (ex. c(.5) for a clonal diploid region, with purity 1)}
}
\value{
the log-likelihood of the set of SNVs
}
\description{
Computes the log-likelihood for a set of SNVs falling on a single segment
}
