% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_peaks.R
\name{analyze_peaks}
\alias{analyze_peaks}
\title{Analyze CNA calls by peak detection.}
\usage{
analyze_peaks(
  x,
  karyotypes = c("1:0", "1:1", "2:1", "2:0", "2:2"),
  min_karyotype_size = 0.05,
  p_binsize_peaks = 0.005,
  matching_epsilon = 0.025,
  kernel_adjust = 1
)
}
\arguments{
\item{x}{An object of class \code{cnaqc}, created by the \code{init} function.}

\item{karyotypes}{The list of karyotypes to test. By default LOH regions (A, AA),
diploid regions (AB), and amplification regions (AAB, AABB). These correspond to
\code{'1:0', '1:1', '2:1', '2:0', '2:2'} in "Major:minor" notation.}

\item{min_karyotype_size}{Karyotypes are subset by their size (normalized for
the number of input mutations). Karyotypes smaller than this value are removed
from analysis.}

\item{p_binsize_peaks}{Peaks detected will be filtered if, in a peak, we map
less than \code{p_binsize_peaks * N} mutations. The value \code{N} is obtained
couting all mutations that map in all peaks.}

\item{matching_epsilon}{Peaks at a location are matched with a range of \code{epsilon * 0.5},
left and right. By default (\code{0.025}) a maxium 5\% tolerance is adopted. The status of
overall PASS/ FAIL for this analysis is determined with the same tolerance, given the score
computed across all karyotypes.}

\item{kernel_adjust}{KDE adjust density parameter; see \code{density}. A Gaussian
kernel is used (\code{kernel = 'gaussian'}).}
}
\value{
An object of class \code{cnaqc}, modified to hold the results from this analysis.
See the vignette to see how to extract and plot the results.
}
\description{
This function carries out a peak-detection
analysis based on KDE and the package \code{peakPick}
in order to determine if the mutations that map to a
certain karyotype have an allelic frequency that is
consistent with the Major and minor alleles of the
segment, and the tumour purity. A score is produced
as a linear combination of the distance of the actual
peak to the expected one, derived with standard equations
for CNA analysis.
}
\examples{
data('example_dataset_CNAqc', package = 'CNAqc')
x = init(example_dataset_CNAqc$snvs, example_dataset_CNAqc$cna,example_dataset_CNAqc$purity)

x = analyze_peaks(x)
print(x)

print(x$peaks_analysis)
}
