% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_peaks.R
\name{analyze_peaks}
\alias{analyze_peaks}
\title{Analyze the peaks of this}
\usage{
analyze_peaks(x, karyotypes = c("1:0", "1:1", "2:1", "2:0", "2:2"),
  min_karyotype_size = 0.05, adjust = 1, min_density = 0.5,
  neighlim = 1, matching_epsilon = 0.015, ...)
}
\arguments{
\item{...}{}
}
\description{
Analyze the peaks of this
}
