% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variants.R
\name{annotate_variants}
\alias{annotate_variants}
\title{Annotate variants and drivers.}
\usage{
annotate_variants(x, drivers = CNAqc::intogen_drivers)
}
\arguments{
\item{x}{A CNAqc object.}

\item{drivers}{A dataframe in the format of the `intogen_drivers` one released with
CNAqc. In particular, it must contain a column named `gene` to identify gene names.}
}
\value{
A CNAqc object with variants annotated. For each variant this object contains:

- `location` reporting the position of the variant in the genome (`coding`, `intron`, `threeUTR`, ...),
- `consequence` with the consequence of coding mutations (`synonymous`, `nonsynonymous`, ...),
- `is_driver` a boolean that indicates if the gene is a driver,
- `gene_symbol` for the annotated corresponding gene symbol (if the variant is in a gene)
- `driver_label` with the driver label written as `gene_refAA->varAA` (`NA` in case `is_driver = FALSE`).

The annotation process is based on the package \href{https://bioconductor.org/packages/release/bioc/html/VariantAnnotation.html}{VariantAnnotation}.
}
\description{
Easy and fast way to annotate input mutations, and detect
potential driver mutations. This function computes the locations of
the different mutations and the consequences of substituions mapped to
coding regions, using \href{https://bioconductor.org/packages/release/bioc/html/VariantAnnotation.html}{VariantAnnotation} and other Bioconductor packages.
Then, putative drivers are annotated upon matching from an input list that,
by default, is compiled from the \href{https://www.intogen.org}{Intogen} database. Drivers are selected among
coding substituions with known effect.
}
\examples{

\dontrun{
library(CNAqc)

data('example_dataset_CNAqc', package = 'CNAqc')

mutations <- example_dataset_CNAqc$mutations

mutations_annotated <- annotate_variants(mutations)
}

}
