% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf.R
\name{as_maftools_cohort}
\alias{as_maftools_cohort}
\title{Convert a list of CNAqc object to a maftools object.}
\usage{
as_maftools_cohort(x, only_drivers = TRUE, CNA_genes = NULL)
}
\arguments{
\item{x}{A list of CNAqc objects with MAF annotations.}

\item{only_drivers}{If `TRUE`, only driver mutations are used, otherwised all.
When `TRUE`, if drivers are not annotated, an error is thrown.}

\item{CNA_genes}{}
}
\value{

}
\description{
Like \code{\link{as_maftools_obj}} but for multiple CNAqc objects, it creates
a unique [maftools](https://bioconductor.org/packages/release/bioc/html/maftools.html)
object with all the samples at once.

Parameters have the same meaning as for function \code{\link{as_maftools_obj}}.
}
\examples{
if(FALSE)
{
   # Create your CNAqc object (omissis here) from an original "file.vcf"
   x = init(mutations = ..., cna = ..., purity = ...)

   # Offline, create your MAF annotations as file "file_vcf.maf" from "file.vcf"
   # vcf2maf file.vcf .... file_vcf.maf

   # Import into R/CNAqc
   x = augment_with_maf(x, maf = "file_vcf.maf")

   # Extraction
   x \%>\% as_maftools_obj
}
}
\seealso{
function \code{\link{augment_with_maf}} to add MAF annotations to a
CNAqc object, to be used before running `as_maftools_obj`.
}
