% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf.R
\name{augment_with_maf}
\alias{augment_with_maf}
\title{Import MAF annotations into a CNAqc object}
\usage{
augment_with_maf(x, maf)
}
\arguments{
\item{x}{A CNAqc object.}

\item{maf}{The file MAF associated containing the annotations in MAF format
for the input object `x`.}
}
\value{
A CNAqc object like `x` where the mutations are associated to the
MAF annotations, matched by genomic coordinates of the mutations. The S3
print method for `x` will report the presence of MAF annotaitons.
}
\description{
This function import MAF annotations into a CNAqc object. MAF annotations
should have been created using the [vcf2maf](https://github.com/mskcc/vcf2maf)
utility. At this point, if one has created the input data for CNAqc from the
original VCF file, the MAF can be added to the CNAqc object. Mutations are
associated based on genome locations and substitutions. Once many CNAqc objects
have been augmented with their MAF annotations, a cohort of MAFs can be
exported and functions from Bioconductor package [maftools](https://bioconductor.org/packages/release/bioc/html/maftools.html)
can be used to plot data from multiple patients.
}
\examples{
if(FALSE)
{
   # Create your CNAqc object (omissis here) from an original "file.vcf"
   x = init(mutations = ..., cna = ..., purity = ...)

   # Offline, create your MAF annotations as file "file_vcf.maf" from "file.vcf"
   # vcf2maf file.vcf .... file_vcf.maf

   # Import into R/CNAqc
   x = augment_with_maf(x, maf = "file_vcf.maf")

   # check they are in (there should be many columns with "MAF." prefix)
   x \%>\% Mutations \%>\% colnames
}
}
\seealso{
function \code{\link{as_maftools_cohort}} to convert multiple CNAqc
objects with MAF annotations into a single MAF cohort;  package
[maftools](https://bioconductor.org/packages/release/bioc/html/maftools.html) to
summarize, analyze and visualize MAF Files; the utility [vcf2maf](https://github.com/mskcc/vcf2maf)
to create MAF files from VCFs, using the [Ensembl Variant Effect Predictor (VEP)](https://www.ensembl.org/info/docs/tools/vep/index.html)
utility.
}
