% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_tolerance.R
\name{auto_tolerance}
\alias{auto_tolerance}
\title{Determine the optimal error tolerance to QC clonal simple  CNAs.}
\usage{
auto_tolerance(purity, coverage, fpr = 0.1, epsilon_range = c(0.01, 0.08))
}
\arguments{
\item{purity}{The data purity (putative) for a sample.}

\item{coverage}{The observed data coverage.}

\item{fpr}{Desired false positive rate.}

\item{epsilon_range}{Range of values to constrain $epsilon$.}
}
\value{
The $\epsilon$ value estimated from data, constrained to be in
`epsilon_range`, in order to limit the false positive rate to be at most `fpr`.
}
\description{
The QC procedure implemented by function \code{analyze_peaks} allows to
pass or fail a sample based on a custom purity error (maximum error to allow).

A regression has been used to measure the rate of false positives
from simulated tumours with variable coverage and purity. This allows to determine
an optimal value of $\epsilon$, the parameter `purity_error` of function
\code{analyze_peaks}, for a desired rate of alse positives basde on the sample
coverage and putative purity.
}
\examples{
# Desired 10\% FPR for a 30\% pure tumour at 90x
auto_tolerance(.3, 90)
}
