% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_tolerance.R
\name{auto_tolerance}
\alias{auto_tolerance}
\title{Title}
\usage{
auto_tolerance(purity, coverage, fpr = 0.1, epsilon_range = c(0.01, 0.08))
}
\arguments{
\item{epsilon_range}{}
}
\value{

}
\description{
Title
}
