% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCF.R
\name{CCF}
\alias{CCF}
\title{Return the available mutation CCF estimates in the data.}
\usage{
CCF(x)
}
\arguments{
\item{x}{An object of class \code{cnaqc}, where CCF have been computed using
function `compute_CCF`.}
}
\value{
A tibble with the mutations with the new columns for mutation multiplicity
and CCF values.
}
\description{
This function extracts Cancer Cell Fractions (CCFs) estimates
from a `CNAqc` object, after they have been computed using
function `compute_CCF`. The estimates are pooled
across the used karyotypes, and extracted from the
`CCF_estimates` field of the input object.
}
\examples{
data('example_dataset_CNAqc')
x = init(example_dataset_CNAqc$snvs, example_dataset_CNAqc$cna, example_dataset_CNAqc$purity)

x = compute_CCF(x)
CCF(x)
}
