% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chr_coordinates_hg19}
\alias{chr_coordinates_hg19}
\title{Coordinates for hg19 chromosomes.}
\format{A tibble that represents the coorinates for hg19 chromosome
(chr, length, from, to, centromerStart and centromerEnd)}
\usage{
data(chr_coordinates_hg19)
}
\description{
Coordinates for hg19 chromosomes.
}
\examples{
data(chr_coordinates_hg19)
chr_coordinates_hg19

}
\references{
hg19
}
\keyword{datasets}
