% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_arm_overfragmentation.R
\name{detect_arm_overfragmentation}
\alias{detect_arm_overfragmentation}
\title{Title}
\usage{
detect_arm_overfragmentation(x, alpha = 0.01,
  genome_percentage_cutoff = 0.2, minimum_segments_for_testing = 10)
}
\arguments{
\item{minimum_segments_for_testing}{}
}
\description{
Title
}
