% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_arm_overfragmentation.R
\name{detect_arm_overfragmentation}
\alias{detect_arm_overfragmentation}
\title{Determines segments' fragmentation status.}
\usage{
detect_arm_overfragmentation(x, alpha = 0.01,
  genome_percentage_cutoff = 0.2, minimum_segments_for_testing = 10)
}
\arguments{
\item{x}{An object of class \code{cnaqc}, created by the \code{init} function.}

\item{alpha}{Confidence level for the tests, for instance \code{0.05}.}

\item{genome_percentage_cutoff}{Segments are considered long or short depending on whether
they are longer (in basepairs) than \code{genome_percentage_cutoff * L} bases, where \code{L}
is the arm length for the reference genome. Default is \code{0.2} (20%).}

\item{minimum_segments_for_testing}{Smallest number of segments required to actually test
a certain arm Default is \code{10} segments. This number influences the correction for mulitple
hypothesis testing.}
}
\value{
An object of class \code{cnaqc} with the results.
}
\description{
This functions determines if, at the arm level, the segments are over-fragmented
using the statistical test described in the package manual.
}
\examples{
data('example_dataset_CNAqc')
x = init(example_dataset_CNAqc$snvs, example_dataset_CNAqc$cna, example_dataset_CNAqc$purity)

x = detect_arm_overfragmentation(x)
print(x)
}
