% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dataset_CNAqc}
\alias{example_dataset_CNAqc}
\title{Example CNAqc dataset.}
\format{
A list of snvs, cna and purity value that
can be used to run CNAqc analyses.
}
\usage{
data(example_dataset_CNAqc)
}
\description{
Example CNAqc dataset.
}
\examples{
data(example_dataset_CNAqc)
example_dataset_CNAqc

}
\references{
??
}
\keyword{datasets}
