% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fpr_test}
\alias{fpr_test}
\title{Data (simulation performance) from the trainig set to auto-tune epsilon.}
\format{
A tibble.
}
\usage{
data(fpr_test)
}
\description{
Data (simulation performance) from the trainig set to auto-tune epsilon.
}
\examples{
data(fpr_test)
fpr_test

}
\keyword{datasets}
