% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_segment.R
\name{inspect_segment}
\alias{inspect_segment}
\title{Plot the VAF of a set of segments}
\usage{
inspect_segment(x, chrs = c(paste0("chr", 1:22), "chrX", "chrY"),
  n = 200, l = 1000)
}
\arguments{
\item{x}{An object of class \code{cnaqc}.}

\item{chrs}{The chromosome ids to use (e.g., "chr2"). All are used by default.}

\item{n}{Disregard chromosome with less than `n` mutations mapped.}

\item{l}{Disregard chromosome that span less than `l` nucleotide.}
}
\value{
A ggplot object of the VAF histogram per segment.
}
\description{
Plot the VAF of a set of segments that map to some
chromosome id, with a minimum number of mapped mutations and length.
THe plot is a histogram of the VAF, coloured per segement. A maximum
of 74 segements can be plotted in colour, otherwise colors are not used
and the histogram is black. The plots are facetted by chromosome id and
karyotype.
}
\examples{
data('example_dataset_CNAqc', package = 'CNAqc')
x = init(example_dataset_CNAqc$snvs, example_dataset_CNAqc$cna,example_dataset_CNAqc$purity)

# Deafault segments -- all chromsomes, at least 1KB, at least 200 mapped mutations.
inspect_segment(x)

# Same as above, but only for chromosome 2
inspect_segment(x, chrs = 'chr2', n = 50)
}
