% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intogen_drivers}
\alias{intogen_drivers}
\title{List of Intogen driver genes per tumour type.}
\format{
A tibble.
}
\usage{
data(intogen_drivers)
}
\description{
List of Intogen driver genes per tumour type.
}
\examples{
data(intogen_drivers)
intogen_drivers

}
\references{
Francisco Martínez-Jiménez, Ferran Muiños, Inés Sentís, Jordi
Deu-Pons, Iker Reyes-Salazar, Claudia Arnedo-Pac, Loris Mularoni, Oriol Pich,
Jose Bonet, Hanna Kranas, Abel Gonzalez-Perez & Nuria Lopez-Bigas.
A compendium of mutational cancer driver genes Nat Rev Cancer (2020).;
doi:10.1038/s41568-020-0290-x
}
\keyword{datasets}
