% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_ccf_purity}
\alias{model_ccf_purity}
\title{Title}
\format{
An object of class \code{list} of length 14.
}
\usage{
data(model_ccf_purity)
}
\description{
Title
}
\examples{
data(model_ccf_purity)
model_ccf_purity

}
\keyword{datasets}
