% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CCF.R
\name{plot_CCF}
\alias{plot_CCF}
\title{Plot the CCF estimates in the data.}
\usage{
plot_CCF(x, strip = FALSE, empty_plot = TRUE, assembly_plot = TRUE)
}
\arguments{
\item{x}{An object of class \code{cnaqc}, where CCF have been computed using
function `compute_CCF`.}

\item{strip}{If \code{TRUE}, shows little detail and a strip horizontal plot.
Otherwise it shows a detailed report (multiple rows). This parameter is now
deprecated.}

\item{empty_plot}{If data for one karyotype is missing, an empty plot is returned.
Otherwise the plot is not returned (NULL is forwarded).}

\item{assembly_plot}{If \code{TRUE}, a unique figure is returned with all the
plots assembled. Otherwise a list of plot is returned.}
}
\value{
A `ggpubr` figure of the CCF estimates.
}
\description{
This function extracts the plots of the Cancer Cell Fractions (CCFs) estimates
from a `CNAqc` object, after they have been computed using
function `compute_CCF`. The plots - one per computed karyotype - are assembled
into a `ggpubr` figure.
}
\examples{
data('example_dataset_CNAqc')
x = init(mutations = example_dataset_CNAqc$mutations, cna =example_dataset_CNAqc$cna, purity = example_dataset_CNAqc$purity)

x = compute_CCF(x)
plot_CCF(x)
}
