% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_histogram.R
\name{plot_data_histogram}
\alias{plot_data_histogram}
\title{Plot the data histogram.}
\usage{
plot_data_histogram(
  x,
  which = "VAF",
  karyotypes = c("1:0", "1:1", "2:0", "2:1", "2:2")
)
}
\arguments{
\item{x}{An object of class \code{cnaqc}, created by the \code{init} function.}

\item{which}{One of \code{"VAF"}, \code{"DP"}, \code{"NV"} or \code{"CCF"}.}

\item{karyotype}{A list of karyotype ids in \code{"Major:minor"} notation
(e.g., \code{"1:1", "2,1", ...}) that will be retained ofr the plot.}
}
\value{
A \code{ggplot} object.
}
\description{
This function plots the histogram of any of the following:
the Variant Allele Frequency (VAF), the depth of sequencing
(DP), the nuber of reads with the variant (NV) and the
Cancer Cell Fractions (CCF) esimates (that must be computed
before.

The plot can be subset by selecting only data in certain segments,
identified by the karyotype shortname, and will be coloured accordingly.
}
\examples{
data('example_dataset_CNAqc', package = 'CNAqc')
x = init(example_dataset_CNAqc$snvs, example_dataset_CNAqc$cna,example_dataset_CNAqc$purity)
plot_data_histogram(x)
plot_data_histogram(x, which = 'DP')
plot_data_histogram(x, which = 'DP', karyotypes = '2:2')

x = compute_CCF(x)
plot_data_histogram(x, which = 'CCF')
}
