% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histograms.R
\name{plot_data_histogram}
\alias{plot_data_histogram}
\title{Title}
\usage{
plot_data_histogram(x, which = "VAF", karyotypes = c("1:0", "1:1",
  "2:0", "2:1", "2:2"))
}
\arguments{
\item{karyotypes}{}
}
\description{
Title
}
