% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome_wide_mutations.R
\name{plot_gw_ccf}
\alias{plot_gw_ccf}
\title{Plot a genome-wide scatter plot of VAF}
\usage{
plot_gw_ccf(x, N = 5000, chromosomes = paste0("chr", c(1:22, "X", "Y")))
}
\arguments{
\item{x}{An object of class \code{cnaqc}, created by the \code{init} function.}

\item{N}{Mutations to use, randomly sampled.}

\item{chromosomes}{The chromosome to use for this plot.}
}
\value{
A \code{ggplot} object.
}
\description{
@description Plot a genome-wide scatter plot of mutation VAFs, downsampled
 if required (annotates the used proportion).
}
\examples{
data('example_dataset_CNAqc', package = 'CNAqc')
x = init(example_dataset_CNAqc$snvs, example_dataset_CNAqc$cna,example_dataset_CNAqc$purity)

# Compute CCF
x = compute_CCF(x)
plot_gw_ccf(x)

plot_gw_ccf(x, N = 100)
plot_gw_ccf(x, N = 1000)
plot_gw_ccf(x, N = 10000)
}
