% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_wide_mutation_plots.R
\name{plot_gw_counts}
\alias{plot_gw_counts}
\title{Plot a genome-wide histogram of mutation counts.}
\usage{
plot_gw_counts(x, chromosomes = paste0("chr", c(1:22, "X", "Y")))
}
\arguments{
\item{x}{An object of class \code{cnaqc}, created by the \code{init} function.}

\item{chromosomes}{The chromosome to use for this plot.}
}
\value{
A \code{ggplot} object.
}
\description{
@description Plot a genome-wide histogram of mutation counts, binned
every one megabase (10e6 positions according to the hg19 reference).
}
\examples{
data('example_dataset_CNAqc', package = 'CNAqc')
x = init(example_dataset_CNAqc$snvs, example_dataset_CNAqc$cna,example_dataset_CNAqc$purity)

plot_gw_counts(x)
}
