% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_icon_cna.R
\name{plot_icon_CNA}
\alias{plot_icon_CNA}
\title{Plot icon for these segments.}
\usage{
plot_icon_CNA(x)
}
\arguments{
\item{x}{A `CNAqc` object.}
}
\value{
A `ggplot` plot.
}
\description{
An icon is a small plot with no legend or axis, which
has the same layout of function \code{plot_multisample_CNA}.
}
\examples{
data('example_dataset_CNAqc', package = 'CNAqc')
x = init(mutations = example_dataset_CNAqc$mutations, cna = example_dataset_CNAqc$cna, purity = example_dataset_CNAqc$purity)

plot_icon_CNA(x)
}
