% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_icon_cna.R
\name{plot_icon_CNA}
\alias{plot_icon_CNA}
\title{A circular plot for simple clonal CNAs.}
\usage{
plot_icon_CNA(x)
}
\arguments{
\item{x}{A CNAqc object.}
}
\value{
A `ggplot2` plot.
}
\description{
An icon is a small circular plot with no legend or axis, which
shows simple clonal CNAs.
}
\examples{
data('example_dataset_CNAqc', package = 'CNAqc')
x = init(mutations = example_dataset_CNAqc$mutations, cna = example_dataset_CNAqc$cna, purity = example_dataset_CNAqc$purity)

plot_icon_CNA(x)
}
