% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_karyotypes.R
\name{plot_karyotypes}
\alias{plot_karyotypes}
\title{Plot counts and numbers of clonal simple CNAs.}
\usage{
plot_karyotypes(
  x,
  type = "percentage",
  chromosomes = paste0("chr", c(1:22, "X", "Y"))
)
}
\arguments{
\item{x}{A CNAqc object.}

\item{type}{With \code{"percentage"}, the proportion of genome covered is returned.
With \code{"number"}, the segment counts. In all other cases an erorr is generated.}

\item{chromosomes}{The chromosome to use for this plot.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Plot a barplot for the segments, reporting either their
 counts of the proportion of genome covered.
}
\examples{
data('example_dataset_CNAqc', package = 'CNAqc')
x = init(mutations = example_dataset_CNAqc$mutations, cna = example_dataset_CNAqc$cna, purity = example_dataset_CNAqc$purity)

# All chromosomes (default)
plot_karyotypes(x)

# Specific chromosomes
plot_karyotypes(x, chromosomes = 'chr3')
plot_karyotypes(x, chromosomes = 'chr13')
}
