% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multisample_CNA.R
\name{plot_multisample_CNA}
\alias{plot_multisample_CNA}
\title{Plots CNA for multiple samples.}
\usage{
plot_multisample_CNA(x)
}
\arguments{
\item{x}{A named list of `CNAqc` objects.}
}
\value{
A `ggplot` plot
}
\description{
This functions plots with circular layout the CNA segments of multiple
`CNAqc` objects, stored in a named list `x`. In this layout, instead of
plotting the Major and minor alleles of each segment, we assign colours
to certain karyotypes. Each sample is plot on a lane, like in a donut plot.
}
\examples{
data('example_dataset_CNAqc', package = 'CNAqc')
x = init(example_dataset_CNAqc$snvs, example_dataset_CNAqc$cna,example_dataset_CNAqc$purity)

plot_multisample_CNA(list(`S1` = x, `S2` = x))
}
