% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_segments.R
\name{plot_segments}
\alias{plot_segments}
\title{Plot CNA segments.}
\usage{
plot_segments(x, chromosomes = paste0("chr", c(1:22, "X", "Y")),
  circular = FALSE)
}
\arguments{
\item{x}{An object of class \code{cnaqc}, created by the \code{init} function.}

\item{chromosomes}{The chromosome to use for this plot.}

\item{circular}{Uses a circular layout in polar coordinates to make the segments
look like a circos plot.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot CNA segments as Major/ minor allele, annotating
clonal and subclonal CNA calls in two different sets of colors.
This function uses \code{hg19} genome coordinates to map CNA segments.
}
\examples{
data('example_dataset_CNAqc', package = 'CNAqc')
x = init(example_dataset_CNAqc$snvs, example_dataset_CNAqc$cna,example_dataset_CNAqc$purity)

plot_segments(x)
plot_segments(x, chromosomes = 'chr13')
}
