% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cnaqc.R
\name{print.cnaqc}
\alias{print.cnaqc}
\title{Print for an object of class \code{'cnaqc'}.}
\usage{
\method{print}{cnaqc}(x, ...)
}
\arguments{
\item{x}{An obj of class \code{'dbpmm'}.}

\item{...}{}
}
\value{
nothing.
}
\description{
Print for an object of class \code{'cnaqc'}.
}
