% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_sbs_contexts.R
\name{SBS_counts}
\alias{SBS_counts}
\title{Extract SBS count data}
\usage{
SBS_counts(x)
}
\arguments{
\item{x}{A CNAqc object.}
}
\value{
A one-row matrix with SBS data for 96 contexts and 6 possible substitutions.
}
\description{
Extract SBS data for 96 contexts and 6 possible substitutions, in matrix
format.
}
