% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsample}
\alias{subsample}
\title{Randomly subsample a dataset}
\usage{
subsample(x, N = 15000, keep_drivers = TRUE)
}
\arguments{
\item{keep_drivers}{}
}
\value{

}
\description{
Randomly subsample a dataset
}
