% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subset_by_minimum_CCF}
\alias{subset_by_minimum_CCF}
\title{Title}
\usage{
subset_by_minimum_CCF(x, min_target_CCF = 0.1)
}
\arguments{
\item{min_target_CCF}{}
}
\value{

}
\description{
Title
}
\examples{

x = init(snvs = example_dataset_CNAqc$snvs, cna = example_dataset_CNAqc$cna, purity = example_dataset_CNAqc$purity)
print(x)

plot_data_histogram(x)

plot_data_histogram(subset_by_minimum_CCF(x))
}
