% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_by_segment.R
\name{subset_by_segment_karyotype}
\alias{subset_by_segment_karyotype}
\title{Subset calls by segments' karyotype.}
\usage{
subset_by_segment_karyotype(x, karyotype)
}
\arguments{
\item{karyotype}{}
}
\description{

}
