% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subset_by_segment_totalcn}
\alias{subset_by_segment_totalcn}
\title{Subset calls by segments total copy number}
\usage{
subset_by_segment_totalcn(x, totalcn)
}
\arguments{
\item{x}{An object of class \code{cnaqc}, created by the \code{init} function.}

\item{total}{A list of copy number integers to find segments to retain only if they
have a certain total copy state.}
}
\value{
An object of class \code{cnaqc}, created by the \code{init} function.
}
\description{
For an object already created, it subsets the calls to those that involve
segments with a certain total copy number (defined as Major plus minor).
}
\examples{
data('example_dataset_CNAqc', package = 'CNAqc')
x = init(example_dataset_CNAqc$mutations, example_dataset_CNAqc$cna,example_dataset_CNAqc$purity)

}
