% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subset_snvs}
\alias{subset_snvs}
\title{Title}
\usage{
subset_snvs(
  x,
  ref_nucleotides = c("A", "C", "T", "G"),
  alt_nucleotides = c("A", "C", "T", "G")
)
}
\arguments{
\item{x}{}

\item{ref_nucleotides}{}

\item{alt_nucleotides}{}
}
\value{

}
\description{
Title
}
